Move notebookCommunication to 'notebooks' folder
