+++ src/notebooks/types.ts
@@ -56,18 +56,6 @@ export enum CellOutputMimeTypes {
     stdout = 'application/vnd.code.notebook.stdout'
 }
 
-/**
- * Handles communications between the WebView (used to render oututs in Notebooks) & extension host.
- */
-export interface INotebookCommunication {
-    readonly editor: NotebookEditor;
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    readonly onDidReceiveMessage: Event<any>;
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    postMessage(message: any): Thenable<boolean>;
-    asWebviewUri(localResource: Uri): Uri;
-}
-
 export const INotebookCompletionProvider = Symbol('INotebookCompletionProvider');
 
 export interface INotebookCompletionProvider {
