+++ src/platform/webviews/types.ts
@@ -1,0 +1,14 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+import { Event, Uri } from 'vscode';
+
+/**
+ * Handles communications between the WebView (used to render oututs in Notebooks) & extension host.
+ */
+export interface IWebviewCommunication {
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
+    readonly onDidReceiveMessage: Event<any>;
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
+    postMessage(message: any): Thenable<boolean>;
+    asWebviewUri(localResource: Uri): Uri;
+}
